pyproject.toml
[project]
name = "mcp-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.1",
    "arxiv>=2.2.0",
    "mcp>=1.9.2",
    "python-dotenv>=1.1.0",
    "typing>=3.10.0.0",
]
